requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Werkzeug==3.0.1
psycopg[binary]==3.3.2
python-dotenv==1.0.1
redis==5.0.7
rq==1.16.2
openai>=1.12.0
gunicorn==21.2.0
Telethon==1.36.0
cryptg==0.4.0
feedparser==6.0.11
beautifulsoup4==4.12.2
python-slugify==8.0.1
requests==2.31.0
Pillow==10.4.0
